Reorganize top level directories into client, server, and common.
Fixes #643
