--- examples/universal/bin/server.js
@@ -1,2 +1,0 @@
-require('babel/register');
-require('../server');
