+++ examples/universal/common/api/counter.js
