+++ examples/universal/common/components/Counter.js
