+++ examples/universal/common/reducers/counter.js
