+++ examples/universal/common/reducers/index.js
