+++ examples/universal/common/store/configureStore.js
